🔢 refactor: use variable type from stdint

+++ strongbox/strongbox.ino
@@ -1,23 +1,26 @@
-//     _____ _                         _                 ____   ___   ___   ___  
+//     _____ _                         _                 ____   ___   ___   ___
 //    / ____| |                       | |               |___ \ / _ \ / _ \ / _ \ 
 //   | (___ | |_ _ __ ___  _ __   __ _| |__   _____  __   __) | | | | | | | | | |
 //    \___ \| __| '__/ _ \| '_ \ / _` | '_ \ / _ \ \/ /  |__ <| | | | | | | | | |
 //    ____) | |_| | | (_) | | | | (_| | |_) | (_) >  <   ___) | |_| | |_| | |_| |
-//   |_____/ \__|_|  \___/|_| |_|\__, |_.__/ \___/_/\_\ |____/ \___/ \___/ \___/ 
-//                                __/ |                                          
-//                               |___/                                           
+//   |_____/ \__|_|  \___/|_| |_|\__, |_.__/ \___/_/\_\ |____/ \___/ \___/ \___/
+//                                __/ |
+//                               |___/
 
 // Version 0.1 - Basique
 
+// Use variable type from stdint
+// https://learn.adafruit.com/deciphering-strange-arduino-code/stdint
+#include <stdint.h>
 
 // Passcode definition
-const int code1 = 1;
-const int code2 = 2;
-const int code3 = 3;
-const int code4 = 4;
+const uint8_t code1 = 1;
+const uint8_t code2 = 2;
+const uint8_t code3 = 3;
+const uint8_t code4 = 4;
 
 // Variables
-int codeStatus;
+uint8_t codeStatus;
 bool shouldError;
 
 void setup() {
@@ -35,35 +38,35 @@ void setup() {
 void loop() {
   // Call buttonPressed and store its output as buttonVar
   // This ensure the button isn't released during the loop execution
-  int buttonVar = buttonPressed();
+  uint8_t buttonVar = buttonPressed();
 
-  if(buttonVar != 0){
-    if(codeStatus == 1){
-      if(buttonVar == code1){
+  if (buttonVar != 0) {
+    if (codeStatus == 1) {
+      if (buttonVar == code1) {
         Serial.println("Passcode authentication - Correct 1/4");
       } else {
         shouldError = true;
         Serial.println("Passcode authentication - Incorrect 1/4");
       }
     }
-    if(codeStatus == 2){
-      if(buttonVar == code2){
+    if (codeStatus == 2) {
+      if (buttonVar == code2) {
         Serial.println("Passcode authentication - Correct 2/4");
       } else {
         shouldError = true;
         Serial.println("Passcode authentication - Incorrect 2/4");
       }
     }
-    if(codeStatus == 3){
-      if(buttonVar == code3){
+    if (codeStatus == 3) {
+      if (buttonVar == code3) {
         Serial.println("Passcode authentication - Correct 3/4");
       } else {
         shouldError = true;
         Serial.println("Passcode authentication - Incorrect 3/4");
       }
     }
-    if(codeStatus == 4){
-      if(buttonVar == code4){
+    if (codeStatus == 4) {
+      if (buttonVar == code4) {
         Serial.println("Passcode authentication - Correct 4/4");
       } else {
         shouldError = true;
@@ -71,11 +74,11 @@ void loop() {
       }
     }
 
-    if(codeStatus != 4){
+    if (codeStatus != 4) {
       codeStatus++;
-      delOutput(codeStatus-1, false);
+      delOutput(codeStatus - 1, false);
     } else {
-      if(shouldError){
+      if (shouldError) {
         Serial.println("Passcode authentication - Failed");
         error();
         codeStatus = 1;
